requirements.txt
flask
psycopg2
sqlalchemy
apscheduler
python-dotenv
flask_sqlalchemy